Rename directory to use precise name
