+++ tool/site/client/tests/components/prescription-summary/practitionerRoleSummaryList.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import {PractitionerRoleSummaryList} from "../../../src/components/prescription/fragments/PractitionerRoleSummaryList"
+import {PractitionerRoleSummaryList} from "../../../src/components/prescription-summary/fragments/PractitionerRoleSummaryList"
 import pretty from "pretty"
 import {summaryPractitionerRole} from "./props"
 import {render} from "@testing-library/react"
