+++ tool/site/client/tests/components/prescription-summary/prescriptionLevelDetail.spec.tsx
@@ -3,7 +3,7 @@ import {render} from "@testing-library/react"
 import pretty from "pretty"
 
 import {prescriptionLevelDetailProps} from "./props"
-import {PrescriptionLevelDetails} from "../../../src/components/prescription/fragments/PrescriptionLevelDetails"
+import {PrescriptionLevelDetails} from "../../../src/components/prescription-summary/fragments/PrescriptionLevelDetails"
 
 test("Renders correctly", async () => {
   const component = <PrescriptionLevelDetails {...prescriptionLevelDetailProps}/>
