+++ tool/site/client/tests/components/prescription-summary/prescriptionSummaryView.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import {PrescriptionSummaryView} from "../../../src/components/prescription/PrescriptionSummaryView"
+import {PrescriptionSummaryView} from "../../../src/components/prescription-summary/PrescriptionSummaryView"
 import {summaryPrescription} from "./props"
 import pretty from "pretty"
 import {render} from "@testing-library/react"
