+++ tool/site/client/tests/components/prescription-summary/props.tsx
@@ -1,11 +1,11 @@
-import {SummaryPractitionerRole} from "../../../src/components/prescription/fragments/PractitionerRoleSummaryList"
-import {SummaryPatient} from "../../../src/components/prescription/fragments/PatientSummaryList"
-import {PrescriptionSummaryViewProps} from "../../../src/components/prescription/PrescriptionSummaryView"
-import {SummaryMedication} from "../../../src/components/prescription/fragments/MedicationSummaryTable"
+import {SummaryPractitionerRole} from "../../../src/components/prescription-summary/fragments/PractitionerRoleSummaryList"
+import {SummaryPatient} from "../../../src/components/prescription-summary/fragments/PatientSummaryList"
+import {PrescriptionSummaryViewProps} from "../../../src/components/prescription-summary/PrescriptionSummaryView"
+import {SummaryMedication} from "../../../src/components/prescription-summary/fragments/MedicationSummaryTable"
 
-import {PrescriptionSummaryProps} from "../../../src/components/prescription/fragments/PrescriptionSummary"
-import {PrescriptionLevelDetailsProps} from "../../../src/components/prescription/fragments/PrescriptionLevelDetails"
-import {EditPrescriptionProps} from "../../../src/components/prescription"
+import {PrescriptionSummaryProps} from "../../../src/components/prescription-summary/fragments/PrescriptionSummary"
+import {PrescriptionLevelDetailsProps} from "../../../src/components/prescription-summary/fragments/PrescriptionLevelDetails"
+import {EditPrescriptionProps} from "../../../src/components/prescription-summary"
 
 export const summaryMedication: SummaryMedication = {
   dispenserNotes: ["See your GP next week", "Don't forget your 5 a day"],
